% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flows.R
\name{addFlows}
\alias{addFlows}
\alias{updateFlows}
\alias{removeFlows}
\alias{clearFlows}
\title{Add or modify flows on a leaflet map}
\usage{
addFlows(map, lng0, lat0, lng1, lat1, color = "blue", weight = 3,
  opacity = 1, dir = NULL, time = NULL, popup = NULL, layerId = NULL,
  timeFormat = NULL, initialTime = NULL)

updateFlows(map, layerId, color = NULL, weight = NULL, opacity = NULL,
  dir = NULL, time = NULL, popup = NULL, timeFormat = NULL,
  initialTime = NULL)

removeFlows(map, layerId)

clearFlows(map)
}
\arguments{
\item{map}{A leaflet map object created with \code{\link[leaflet]{leaflet}}.}

\item{lng0}{Longitude of the origin of the flow.}

\item{lat0}{Lattitude of the origin of the flow.}

\item{lng1}{Longitude of the destination of the flow.}

\item{lat1}{Lattitude of the destination of the flow.}

\item{color}{Color of the flow.}

\item{weight}{Size of the flow (in pixels).}

\item{opacity}{Opacity of the flow.}

\item{dir}{Direction of the flow. 1 indicates that the flow goes from origin
to destination and -1 indicates that it goes from destination to origin. If
0, the arrow is not drawn. If \code{NULL}, then it is equal to the sign of
\code{weight}.}

\item{time}{A vector with length equal to the number of rows in \code{chartdata}
and containing either numbers representing time indices or dates or
datetimes. Each unique value must appear as many times as the others. This
parameter can be used when one wants to represent the evolution of some
variables on a map.}

\item{popup}{Content of the popup bind to a given chart. This can be html
text.}

\item{layerId}{An ID variable. It is mandatoy when one wants to update some
chart with \code{updateMinicharts}.}

\item{timeFormat}{Character string used to format dates and times when
argument \code{time} is a \code{Date}, \code{POSIXct} or \code{POSIXlt}
object. See \code{\link[base]{strptime}} for more information.}

\item{initialTime}{This argument can be used to set the initial time step
shown when the map is created. It is used only when argument \code{time} is
set.}
}
\value{
The modified leaflet map object.
}
\description{
These functions can be used to represent flows and their evolution on a map
created with \code{\link[leaflet]{leaflet}()}. Flows are simply represented
by a segment between two points with an arrow at its center that indicates the
direction of the flow.
}
