% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{d3.schemeCategory10}
\alias{d3.schemeCategory10}
\title{d3 color palette}
\format{An object of class \code{character} of length 10.}
\usage{
d3.schemeCategory10
}
\description{
A character vector containing ten colors. These colors are used as the default
color palette
}
\references{
\url{https://github.com/d3/d3-scale}
}
\author{
Francois Guillem \email{francois.guillem@rte-france.com}
}
\keyword{datasets}
