% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{regions}
\alias{regions}
\title{Contours of the french new regions}
\format{An object of class \code{SpatialPolygons} of length 13.}
\usage{
regions
}
\description{
An object of class \code{SpatialPolygons} containing the contours of the
french new regions.
}
\references{
\url{http://www.rte-france.com/fr/eco2mix/eco2mix}
}
\author{
Francois Guillem \email{francois.guillem@rte-france.com}
}
\keyword{datasets}
