% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{eco2mix}
\alias{eco2mix}
\alias{eco2mixBalance}
\title{Electric production, consumption and exchanges of France}
\format{An object of class \code{data.frame} with 686 rows and 22 columns.}
\usage{
eco2mix

eco2mixBalance
}
\description{
\code{eco2mix} contains the electric production, consumption and exchanges
of France from january 2010 to february 2017 and of 12 french regions from
january 2013 to february 2017.

In addition to the total production, the table contains one column for each
type of production. The table also contains the lattitude and longitude of
the center of the regions.

\code{eco2mixBalance} is an extract of \code{eco2mix} that contains only
exchanges between France and neighbouring countries, in a convenient format
to represent flows on a map.
}
\references{
\url{http://www.rte-france.com/fr/eco2mix/eco2mix}
}
\author{
Francois Guillem \email{francois.guillem@rte-france.com}
}
\keyword{datasets}
